requirements.txt
gitpython
PyGithub
tqdm
click